backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
pydantic-settings==2.5.2
google-genai==1.1.0
httpx==0.27.2
python-multipart==0.0.12
reportlab==4.2.5
python-dotenv==1.0.1
Pillow==10.4.0
jinja2==3.1.4
